Refactor two functions into utils -file.
